+++ tools/__init__.py
@@ -1,0 +1,1 @@
+"""Maintainer tooling for rack-preset-kit."""